bug web manual unicode

+++ sys/MelderThread.h
@@ -121,21 +121,21 @@ inline static int MelderThread_getNumberOfProcessors () {
 		if (numberOfThreads == 1) {
 			func (args [0].get());
 		} else {
-			std::vector <pthread_t> threads (numberOfThreads);
+			std::vector <pthread_t> threads ((size_t) numberOfThreads);
 			try {
 				for (int ithread = 1; ithread < numberOfThreads; ithread ++) {
-					(void) pthread_create (& threads [ithread - 1],
+					(void) pthread_create (& threads [(size_t) ithread - 1],
 						nullptr, (void*(*)(void *)) func, (void *) args [ithread - 1].get());
 				}
 				func (args [numberOfThreads - 1].get());
 			} catch (MelderError) {
 				for (int ithread = 1; ithread < numberOfThreads; ithread ++) {
-					pthread_join (threads [ithread - 1], nullptr);
+					pthread_join (threads [(size_t) ithread - 1], nullptr);
 				}
 				throw;
 			}
 			for (int ithread = 1; ithread < numberOfThreads; ithread ++) {
-				pthread_join (threads [ithread - 1], nullptr);
+				pthread_join (threads [(size_t) ithread - 1], nullptr);
 			}
 		}
 	}
